mm: Dynamic management of dirty page writebacks
Originally developed By: XDA Developer Christopher83
Modified/Merged for MTK (Mediatek) By: XDA Developer superdragonpt

This feature allows to dynamically manage the dirty page writebacks with two different intervals, one when the screen is on and another when the screen is off.
It is based on a commit of Francisco Franco, but instead of using hard coded values and of disabling at all the dirty page writebacks while the screen is on (possibility of data loss), I preferred to use a customizable higher dirty page writebacks interval (15 seconds) than the default one (5 seconds) while the screen is on and a customizable default interval (5 seconds) when the screen is off.
By using a higher interval we have better performance and less battery consumption, with a very low risk of data loss.

Change-Id: If2260ffa3ed1599f21dcb00936f8b56048389207